refactor(feature_iptv): extract platform-agnostic domain logic to feature_iptv_core
Phase 2 of the SSOT migration (ADR-0011): move the 7 domain files with
no Flutter-widget or per-shell dependencies (navigation/settings
manifests, local search, VOD grouping/resume, favorite reimport,
wakelock debouncing) into a new feature_iptv_core package.

feature_iptv's old lib/domain/*.dart paths become one-line re-export
shims, so every existing consumer (relative imports within the
package, package:feature_iptv/domain/* imports from tests) keeps
working unchanged. No public API or behavior changes.

This is the first slice of splitting feature_iptv into core/mobile/tv
adapters per the phase 2 plan in
tasks/ssot_airo_airo_tv_architecture_blueprint.md.
